checked post exist or not while deleting post

+++ src/main/java/edu/miu/cs540/eaprojects/controller/PostCommentController.java
@@ -48,8 +48,8 @@ public class PostCommentController {
 
     @DeleteMapping("/comments/delete-all-by-post-id/{postId}")
     public ResponseEntity deleteAllCommentByPostId(@PathVariable Long postId) {
-//        PostDto postDto = postProxyClient.getPostById(postId);
-//        if (postDto == null) throw new PostNotFoundException("Post not found");
+        PostDto postDto = postProxyClient.getPostById(postId);
+        if (postDto == null) throw new PostNotFoundException("Post not found");
         postCommentService.deleteAllByPostId(postId);
         return ResponseEntity.ok().build();
